# backend/src/config/openapi.generated.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Signet Workforce ERP API",
    "version": "1.0.0",
    "description": "Production REST API for Signet Workforce ERP. Authenticate via POST /api/auth/login, then use the Bearer token for protected routes."
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/leave": {
      "get": {
        "tags": [
          "Leave"
        ],
        "summary": "List leave requests",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated leave requests"
          }
        }
      }
    },
    "/api/leave/request": {
      "post": {
        "tags": [
          "Leave"
        ],
        "summary": "Submit leave request",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "employeeId",
                  "leaveType",
                  "fromDate",
                  "toDate",
                  "reason"
                ],
                "properties": {
                  "employeeId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "leaveType": {
                    "type": "integer"
                  },
                  "fromDate": {
                    "type": "string",
                    "format": "date"
                  },
                  "toDate": {
                    "type": "string",
                    "format": "date"
                  },
                  "isHalfDay": {
                    "type": "boolean"
                  },
                  "reason": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Leave request created"
          }
        }
      }
    },
    "/api/leave/{id}/approve": {
      "put": {
        "tags": [
          "Leave"
        ],
        "summary": "Approve or reject leave request",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "approve"
                ],
                "properties": {
                  "approve": {
                    "type": "boolean"
                  },
                  "comments": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Leave status updated"
          }
        }
      }
    },
...
    "/api/notifications": {
      "get": {
        "tags": [
          "Notifications"
        ],
        "summary": "List notifications for current user",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "pageSize",
            "schema": {
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "unreadOnly",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated notifications"
          }
        }
      }
    },
...
    "/api/notifications/{id}/read": {
      "put": {
        "tags": [
          "Notifications"
        ],
        "summary": "Mark one notification as read",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Marked read"
          }
        }
      }
    },
...
  },
...
}